pyproject.toml
[project]
name = "theoremforge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<=3.13"
dependencies = [
    "aiohttp>=3.13.1",
    "datasets>=4.2.0",
    "google-genai>=1.46.0",
    "lean-interact>=0.9.1",
    "loguru>=0.7.3",
    "pyyaml>=6.0.3",
    "uvicorn>=0.38.0",
    "vllm>=0.11.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=7.0.1",
]

